Post
** create a post

In this process, I have been working on the post controller, model, route and also validation. The user must have an account and login in order to create a post.

- Since the database is already created, now in my controller, I added a function that will allow a user to create a post if all the validations are passed.
- In my post router, I created an endpoint that will allow the user to create a post.
- In my validation, i created the different post validation using joi

** Get all the posts

In this process, I have been working on the post controller, model, route and also validation. Anyone can get all the posts.

- Since the database is already created, now in my controller, I added a function that will allow a user to get all posts if all the validations are passed.
- In my post router, I created an endpoint that will allow the user to get all the posts.

** Get a post

In this process, I have been working on the post controller, model, route and also validation. Anyone can get all the posts.

- Since the database is already created, now in my controller, I added a function that will allow a user to get a post if all the validations are passed.
- In my post router, I created an endpoint that will allow the user to get a post.

** Delete a post
In this process, I have been working on the post controller, model, route and also validation. The user must have an account and login in order to delete a post.

- Since the database is already created, now in my controller, I added a function that will allow a user to dele a post only their post if all the validations are passed.
- In my user router, I created an endpoint that will allow the user to delete the post.

** Update a post

- now in my controller, I added a function that will allow a user to dele a post only their post if all the validations are passed.
- In this process, I have been working on the post controller, model, route and also validation. Anyone can get a posts.